Extract read_target_helpers.rs from turn.rs (parent #680)
Move the read-tool path lookup helpers out of `turn.rs` into a focused
sibling module so `turn.rs` keeps shrinking toward responsibility-focused
units (parent meta-issue #680).

Two free fn helpers extracted (no Agent dependency — pure projections
over message slices):

- `last_read_tool_path` — most recent (across all turns) `Read`
  tool-call `path` argument.
- `latest_turn_preferred_read_edit_target` — latest user turn's
  `Read` calls, resolved + filtered to existing files; prefers
  `is_preferred_read_edit_target` matches, otherwise falls back to
  the most-recent existing read.

Behavior unchanged: same message iteration order (reverse for
`last_read_tool_path`, `latest_user_turn_slice` for the per-turn
helper), same `resolve_user_path` workspace confinement, same
`is_preferred_read_edit_target` priority.

`pub(super)` limited / no facade re-export (DR3-001). Callers updated
in `recovery_targets.rs`, `forced_small_edit.rs`, `tool_prep.rs`,
`scaffold_pipeline.rs`, and `progress_tests.rs` (5 import paths).

LOC delta:
- turn.rs: 229 → 175 (-54)
- new module: +76

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
