+++ src/agent/loop_run/tool_prep.rs
@@ -21,9 +21,9 @@
 use std::path::PathBuf;
 
 use super::Agent;
+use super::read_target_helpers::latest_turn_preferred_read_edit_target;
 use super::tool_history::{focused_edit_target_already_read, has_successful_non_plan_repo_edit};
 use super::tool_policy::EffectiveToolPolicy;
-use super::turn::latest_turn_preferred_read_edit_target;
 use crate::modes::plan_act::{ExecutionMode, WorkMode};
 use crate::session::store::ConversationMessage;
 use crate::tools::registry::ToolSpec;
