+++ src/agent/loop_run/turn.rs
@@ -1,9 +1,8 @@
 use super::small_helpers::raw_mode_safe_text;
-use super::tool_history::{is_preferred_read_edit_target, latest_user_turn_slice};
 use super::workspace_walk::workspace_appears_empty;
 use super::*;
 use crate::session::store::ScaffoldArtifactFileSnapshot;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use super::quality::repo_change_request_text;
 
@@ -173,57 +172,4 @@ impl Agent {
             .messages
             .push(ConversationMessage::user(content));
     }
-}
-
-pub(super) fn last_read_tool_path(messages: &[ConversationMessage]) -> Option<String> {
-    messages.iter().rev().find_map(|message| {
-        if message.role != "assistant" {
-            return None;
-        }
-        message.tool_calls.iter().rev().find_map(|tool_call| {
-            if tool_call.name != "Read" {
-                return None;
-            }
-            tool_call
-                .arguments
-                .get("path")
-                .and_then(serde_json::Value::as_str)
-                .map(ToString::to_string)
-        })
-    })
-}
-
-pub(super) fn latest_turn_preferred_read_edit_target(
-    messages: &[ConversationMessage],
-    work_root: &Path,
-) -> Option<PathBuf> {
-    let mut latest_existing = None;
-    for message in latest_user_turn_slice(messages).iter().rev() {
-        if message.role != "assistant" {
-            continue;
-        }
-        for tool_call in message.tool_calls.iter().rev() {
-            if tool_call.name != "Read" {
-                continue;
-            }
-            let Some(path) = tool_call
-                .arguments
-                .get("path")
-                .and_then(serde_json::Value::as_str)
-            else {
-                continue;
-            };
-            let Ok(candidate) = resolve_user_path(work_root, path) else {
-                continue;
-            };
-            if !candidate.is_file() {
-                continue;
-            }
-            latest_existing.get_or_insert_with(|| candidate.clone());
-            if is_preferred_read_edit_target(&candidate) {
-                return Some(candidate);
-            }
-        }
-    }
-    latest_existing
 }
